#79 [Feat] 핵심경험 컴포넌트 분리
